refactor: organize and optimization
