+++ src/app/constants/storageKeys.ts
@@ -1,0 +1,6 @@
+export const storageKeys = {
+  1: 'promptpayID',
+  2: 'promptpayID2',
+  3: 'promptpayID3',
+  4: 'promptpayID4',
+}
