+++ src/app/services/sanitizeId.ts
@@ -1,0 +1,3 @@
+export const sanitizeId = (id: string) => {
+  return String(id).replace(/[^0-9]/g, '')
+}
