+++ src/modules/flipper/@types/FlipperModel.ts
@@ -1,6 +1,6 @@
-export interface IFlipperModel {
+export interface FlipperModel {
   setFlipped(flipped: boolean): void
   pointerDown(): void
   pointerMove: (delta: number) => void
-  pointerUp(): void;
+  pointerUp(): void
 }
