+++ src/modules/localize/components/LocalizedText.tsx
@@ -1,0 +1,7 @@
+import { useLocalization } from '../services/useLocalization'
+
+export function LocalizedText(props: { th: string; en: string }) {
+  const { main, sub } = useLocalization(props.th, props.en)
+
+  return <span title={sub}>{main}</span>
+}
