+++ src/modules/localize/context/localizationContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'preact'
+
+export const localizationContext = createContext('th')
+export const LocalizationProvider = localizationContext.Provider
