+++ src/modules/localize/services/t.tsx
@@ -1,0 +1,5 @@
+import { LocalizedText } from '../components/LocalizedText'
+
+export function t(th: string, en: string) {
+  return <LocalizedText th={th} en={en} />
+}
