+++ src/modules/qrcode/components/QRCode.storybook.js
