+++ src/modules/qrcode/services/createExplosionRenderer.ts
