+++ src/modules/qrcode/services/createPixelsRenderer.ts
