+++ src/modules/slotSelector/styles/SlotSelector.css
